requirements.txt
langchain
langchain-openai
langchain-ollama
langgraph
yfinance
requests
python-dotenv
